@@ -8,17 +8,15 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "bref/bref": "^0.5.31",
-        "bref/laravel-bridge": "^0.2.2",
+        "php": "^7.3.0",
         "fakerphp/faker": "^1.12",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "^7.0.1",
         "laravel/cashier": "^12.1",
-        "laravel/framework": "^7.0",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
+        "laravel/ui": "^3.0",
         "nicolaslopezj/searchable": "^1.13",
         "predis/predis": "^1.1",
         "proj4php/proj4php": "^2.0",
@@ -26,10 +24,10 @@
         "spatie/laravel-image-optimizer": "^1.6"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "mockery/mockery": "^1.3.1",
         "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -46,12 +44,10 @@
             "app/Helpers/Helper.php"
         ],
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
