@@ -1,5 +1,8 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\Review;
 use App\Spot;
 use App\User;
 use Illuminate\Database\Seeder;
@@ -7,7 +10,7 @@ use Illuminate\Database\Seeder;
 class ReviewSeeder extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Run the database seeders.
      *
      * @return void
      */
@@ -15,7 +18,7 @@ class ReviewSeeder extends Seeder
     {
         foreach (User::pluck('id') as $user) {
             foreach (Spot::inRandomOrder()->limit(25)->get() as $spot) {
-                factory(App\Review::class)->create(['spot_id' => $spot->id, 'user_id' => $user]);
+                Review::factory()->create(['spot_id' => $spot->id, 'user_id' => $user]);
 
                 $spot->rating = round($spot->reviews->sum('rating') / count($spot->reviews));
                 $spot->save();
