@@ -1,19 +1,22 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\Spot;
 use App\User;
 use Illuminate\Database\Seeder;
 
 class SpotSeeder extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Run the database seeders.
      *
      * @return void
      */
     public function run()
     {
         foreach (User::pluck('id') as $user) {
-            factory(App\Spot::class, rand(0, 25))->create(['user_id' => $user]);
+            Spot::factory()->times(rand(0, 25))->create(['user_id' => $user]);
         }
     }
 }
