@@ -1,21 +1,24 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\User;
 use Illuminate\Database\Seeder;
 
 class UserSeeder extends Seeder
 {
     /**
-     * Run the database seeds.
+     * Run the database seeders.
      *
      * @return void
      */
     public function run()
     {
-        factory(App\User::class)->create([
+        User::factory()->create([
             'name' => 'Admin',
             'email' => 'admin@parkourhub.com',
             'remember_token' => 'Hfkk1ZgTEAnGZHQdmyJQ8BNU761i94m0yzyHIgZFi6Ir0BLHotRoDFUjLGKA',
         ]);
-        factory(App\User::class, 5)->create();
+        User::factory()->times(5)->create();
     }
 }
