@@ -3,6 +3,7 @@
 namespace App;
 
 use App\Traits\Reportable;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Facades\Auth;
@@ -12,7 +13,8 @@ class Challenge extends Model
 {
     use SoftDeletes,
         Reportable,
-        SearchableTrait;
+        SearchableTrait,
+        HasFactory;
 
     protected $fillable = [
         'name',
