@@ -3,6 +3,7 @@
 namespace App;
 
 use Illuminate\Contracts\Auth\MustVerifyEmail;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -14,7 +15,8 @@ class User extends Authenticatable implements MustVerifyEmail
     use Notifiable,
         SoftDeletes,
         Billable,
-        SearchableTrait;
+        SearchableTrait,
+        HasFactory;
 
     /**
      * The attributes that are mass assignable.
