@@ -1,16 +1,22 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Challenge;
 use App\ChallengeEntry;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(ChallengeEntry::class, function (Faker $faker) {
-    return [
-        'challenge_id' => Challenge::inRandomOrder()->first()->id,
-        'user_id' => User::inRandomOrder()->first()->id,
-        'youtube' => 'Oykjn35X3EY',
-    ];
-});
+class ChallengeEntryFactory extends Factory
+{
+    protected $model = ChallengeEntry::class;
+
+    public function definition()
+    {
+        return [
+            'challenge_id' => Challenge::inRandomOrder()->first()->id,
+            'user_id' => User::inRandomOrder()->first()->id,
+            'youtube' => 'Oykjn35X3EY',
+        ];
+    }
+}
