@@ -1,21 +1,27 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Challenge;
 use App\Spot;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Challenge::class, function (Faker $faker) {
-    return [
-        'spot_id' => Spot::inRandomOrder()->first()->id,
-        'user_id' => User::inRandomOrder()->first()->id,
-        'name' => $faker->word,
-        'description' => $faker->realText(255),
-        'difficulty' => floor($faker->numberBetween(1, 5)),
-        'visibility' => $faker->randomElement(['private', 'follower', 'public']),
-        'youtube' => 'Oykjn35X3EY',
-        'thumbnail' => '/storage/images/spots/' . $faker->image('public/storage/images/spots', 640, 480, null, false),
-    ];
-});
+class ChallengeFactory extends Factory
+{
+    protected $model = Challenge::class;
+
+    public function definition()
+    {
+        return [
+            'spot_id' => Spot::inRandomOrder()->first()->id,
+            'user_id' => User::inRandomOrder()->first()->id,
+            'name' => $this->faker->word,
+            'description' => $this->faker->realText(255),
+            'difficulty' => floor($this->faker->numberBetween(1, 5)),
+            'visibility' => $this->faker->randomElement(['private', 'follower', 'public']),
+            'youtube' => 'Oykjn35X3EY',
+            'thumbnail' => $this->faker->image('public/storage/images/challenges', 640, 480, null, true),
+        ];
+    }
+}
