@@ -1,18 +1,24 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Review;
 use App\Spot;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Review::class, function (Faker $faker) {
-    return [
-        'spot_id' => Spot::inRandomOrder()->first()->id,
-        'user_id' => User::inRandomOrder()->first()->id,
-        'rating' => floor($faker->numberBetween(1, 5)),
-        'title' => $faker->word,
-        'review' => $faker->realText(255),
-    ];
-});
+class ReviewFactory extends Factory
+{
+    protected $model = Review::class;
+
+    public function definition()
+    {
+        return [
+            'spot_id' => Spot::inRandomOrder()->first()->id,
+            'user_id' => User::inRandomOrder()->first()->id,
+            'rating' => floor($this->faker->numberBetween(1, 5)),
+            'title' => $this->faker->word,
+            'review' => $this->faker->realText(255),
+        ];
+    }
+}
