@@ -1,32 +1,38 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Spot;
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use proj4php\Point;
 use proj4php\Proj;
 use proj4php\Proj4php;
 
-$factory->define(Spot::class, function (Faker $faker) {
-    $proj4 = new Proj4php();
-    $proj4326 = new Proj('EPSG:4326', $proj4);
-    $proj3857 = new Proj('EPSG:3857', $proj4);
+class SpotFactory extends Factory
+{
+    protected $model = Spot::class;
 
-    $lat = $faker->latitude;
-    $lon = $faker->longitude;
-    $lonlat = new Point($lon, $lat, $proj4326);
-    $coordinates = $proj4->transform($proj3857, $lonlat);
+    public function definition()
+    {
+        $proj4 = new Proj4php();
+        $proj4326 = new Proj('EPSG:4326', $proj4);
+        $proj3857 = new Proj('EPSG:3857', $proj4);
 
-    return [
-        'user_id' => User::inRandomOrder()->first()->id,
-        'name' => $faker->word,
-        'description' => $faker->realText(255),
-        'visibility' => $faker->randomElement(['private', 'follower', 'public']),
-        'coordinates' => $coordinates->x . ',' . $coordinates->y,
-        'latitude' => $lat,
-        'longitude' => $lon,
-        'image' => '/storage/images/spots/' . $faker->image('public/storage/images/spots', 640, 480, null, false),
-    ];
-});
+        $lat = $this->faker->latitude;
+        $lon = $this->faker->longitude;
+        $lonlat = new Point($lon, $lat, $proj4326);
+        $coordinates = $proj4->transform($proj3857, $lonlat);
+
+        return [
+            'user_id' => User::inRandomOrder()->first()->id,
+            'name' => $this->faker->word,
+            'description' => $this->faker->realText(255),
+            'visibility' => $this->faker->randomElement(['private', 'follower', 'public']),
+            'coordinates' => $coordinates->x . ',' . $coordinates->y,
+            'latitude' => $lat,
+            'longitude' => $lon,
+            'image' => $this->faker->image('public/storage/images/spots', 640, 480, null, true),
+        ];
+    }
+}
