@@ -1,31 +1,26 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+class UserFactory extends Factory
+{
+    protected $model = User::class;
 
-$factory->define(User::class, function (Faker $faker) {
-    return [
-        'name' => $faker->firstName . ' ' . $faker->lastName,
-        'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
-        'profile_image' => '/storage/images/users/profile/' . $faker->image('public/storage/images/users/profile', 640, 480, 'people', false),
-        'cover_image' => '/storage/images/users/cover/' . $faker->image('public/storage/images/users/cover', 640, 480, 'nature', false),
-        'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // 'password'
-        'remember_token' => Str::random(10),
-        'settings' => '{}',
-    ];
-});
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->firstName . ' ' . $this->faker->lastName,
+            'email' => $this->faker->unique()->safeEmail,
+            'email_verified_at' => now(),
+            'profile_image' => $this->faker->image('public/storage/images/users/profile/', 640, 480, null, true),
+            'cover_image' => $this->faker->image('public/storage/images/users/cover/', 640, 480, null, true),
+            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // 'password'
+            'remember_token' => Str::random(10),
+            'settings' => '{}',
+        ];
+    }
+}
